+++ tests/cli_test.py
@@ -25,6 +25,7 @@ structured_app_dir = this_dir / "structured_app"
 @pytest.mark.parametrize(
     "args",
     [
+        # f"{structured_app_dir} --stop-on-error --regen-schemas -vvv",
         f"{this_dir} --stop-on-error --regen-schemas -vvv",
         # TODO: Perhaps we could try to import the repo_root as a python module, or add it to
         # path, or something similar, so that we can actually get a "hydrated" ConfigStore object?
